+++ src/approval/getCommitsToDestination.ts
@@ -3,7 +3,7 @@ import { PullsListCommitsResponseItem } from '../types';
 function commitMatches(commit: string, match: string): boolean {
   // using the length of the proposed match allows for matching shas of
   // different length, so that abcdefghij12345678 matches abcdefghij
-  return commit.slice(0, match.length) === match;
+  return !!match?.length && commit.slice(0, match.length) === match;
 }
 
 export default function getCommitsToDestination(