+++ src/steps/index.ts
@@ -1,4 +1,6 @@
 import { accountSteps } from './account';
+import { collaboratorSteps } from './collaborators';
+import { appSteps } from './apps';
 import { memberSteps } from './members';
 import { prSteps } from './pullrequests';
 import { repoSteps } from './repos';
@@ -9,7 +11,9 @@ const integrationSteps = [
   ...memberSteps,
   ...repoSteps,
   ...teamSteps,
+  ...collaboratorSteps,
   ...prSteps,
+  ...appSteps,
 ];
 
 export { integrationSteps };