+++ src/steps/members.ts
@@ -18,6 +18,8 @@ import {
   GITHUB_MEMBER_ENTITY_CLASS,
   GITHUB_MEMBER_ACCOUNT_RELATIONSHIP_TYPE,
   GITHUB_ACCOUNT_MEMBER_RELATIONSHIP_TYPE,
+  GITHUB_MEMBER_ARRAY,
+  GITHUB_MEMBER_BY_LOGIN_MAP,
 } from '../constants';
 
 export async function fetchMembers({
@@ -68,8 +70,8 @@ export async function fetchMembers({
     }
   });
 
-  await jobState.setData('MEMBER_ARRAY', memberEntities);
-  await jobState.setData('MEMBER_BY_LOGIN_MAP', memberByLoginMap);
+  await jobState.setData(GITHUB_MEMBER_ARRAY, memberEntities);
+  await jobState.setData(GITHUB_MEMBER_BY_LOGIN_MAP, memberByLoginMap);
 }
 
 export const memberSteps: IntegrationStep<IntegrationConfig>[] = [