+++ src/sync/converters.ts
@@ -1,35 +1,13 @@
 import {
   setRawData,
   parseTimePropertyValue,
+  Relationship,
+  RelationshipClass,
 } from '@jupiterone/integration-sdk-core';
 
 import {
-  AccountEntity,
-  RepoEntity,
-  UserEntity,
-  PullRequestEntity,
-  PRState,
-  IdEntityMap,
-  TeamEntity,
-  AccountType,
-  PullsListResponseItem,
-  PullsListCommitsResponseItem,
-} from '../types';
-import { Approval } from '../approval/collectCommitsForPR';
-import {
-  aggregateProperties,
-  flattenMatrix,
-  displayNamesFromLogins,
-} from '../util/propertyHelpers';
-import {
-  OrgMemberQueryResponse,
-  OrgRepoQueryResponse,
-  OrgTeamQueryResponse,
-  OrgQueryResponse,
-  OrgTeamMemberQueryResponse,
-} from '../client/GraphQLClient';
-
-import {
+  GITHUB_REPO_TEAM_RELATIONSHIP_TYPE,
+  GITHUB_REPO_USER_RELATIONSHIP_TYPE,
   GITHUB_ACCOUNT_ENTITY_TYPE,
   GITHUB_ACCOUNT_ENTITY_CLASS,
   GITHUB_MEMBER_ENTITY_TYPE,
@@ -40,7 +18,47 @@ import {
   GITHUB_PR_ENTITY_CLASS,
   GITHUB_TEAM_ENTITY_TYPE,
   GITHUB_TEAM_ENTITY_CLASS,
+  GITHUB_COLLABORATOR_ENTITY_CLASS,
+  GITHUB_COLLABORATOR_ENTITY_TYPE,
+  GITHUB_APP_ENTITY_CLASS,
+  GITHUB_APP_ENTITY_TYPE,
 } from '../constants';
+
+import {
+  AccountEntity,
+  AppEntity,
+  RepoEntity,
+  UserEntity,
+  PullRequestEntity,
+  PRState,
+  IdEntityMap,
+  TeamEntity,
+  AccountType,
+  PullsListResponseItem,
+  PullsListCommitsResponseItem,
+  RepoTeamRelationship,
+} from '../types';
+import { Approval } from '../approval/collectCommitsForPR';
+import {
+  aggregateProperties,
+  flattenMatrix,
+  displayNamesFromLogins,
+  decomposePermissions,
+  getAppEntityKey,
+} from '../util/propertyHelpers';
+import {
+  OrgMemberQueryResponse,
+  OrgRepoQueryResponse,
+  OrgTeamQueryResponse,
+  OrgQueryResponse,
+  OrgTeamMemberQueryResponse,
+  OrgCollaboratorQueryResponse,
+  CollaboratorPermissions,
+  OrgAppQueryResponse,
+} from '../client/GraphQLClient';
+
+import uniq from 'lodash.uniq';
+import omit from 'lodash.omit';
 
 export function toAccountEntity(data: OrgQueryResponse): AccountEntity {
   const accountEntity: AccountEntity = {
@@ -57,13 +75,42 @@ export function toAccountEntity(data: OrgQueryResponse): AccountEntity {
   return accountEntity;
 }
 
+export function toAppEntity(data: OrgAppQueryResponse): AppEntity {
+  const appEntity: AppEntity = {
+    _class: [GITHUB_APP_ENTITY_CLASS],
+    _type: GITHUB_APP_ENTITY_TYPE,
+    _key: getAppEntityKey(data.id),
+    name: data.app_slug,
+    displayName: data.app_slug,
+    webLink: data.html_url,
+    installationId: data.id, //the installation id
+    respositorySelection: data.respository_selection,
+    appId: data.app_id,
+    appSlug: data.app_slug,
+    targetId: data.target_id,
+    targetType: data.target_type,
+    createdAt: data.created_at,
+    updatedAt: data.updated_at,
+    events: data.events,
+    repositorySelection: data.respository_selection,
+    singleFileName: data.single_file_name || '',
+    hasMultipleSingleFiles: data.has_multiple_single_files,
+    singleFilePaths: data.single_file_paths,
+    suspendedBy: data.suspended_by || '',
+    suspendedAt: data.suspended_at || '',
+    ...decomposePermissions(data.permissions),
+  };
+  setRawData(appEntity, { name: 'default', rawData: data });
+  return appEntity;
+}
+
 export function toTeamEntity(data: OrgTeamQueryResponse): TeamEntity {
   const teamEntity: TeamEntity = {
     _class: [GITHUB_TEAM_ENTITY_CLASS],
     _type: GITHUB_TEAM_ENTITY_TYPE,
     _key: data.id,
     webLink: data.url,
-    name: data.slug, //this works, but why? Where is .slug set?
+    name: data.slug,
     displayName: data.name,
     fullName: data.name,
   };
@@ -127,6 +174,25 @@ export function toOrganizationMemberEntityFromTeamMember(
   return userEntity;
 }
 
+export function toOrganizationCollaboratorEntity(
+  data: OrgCollaboratorQueryResponse,
+): UserEntity {
+  const userEntity: UserEntity = {
+    _class: [GITHUB_COLLABORATOR_ENTITY_CLASS],
+    _type: GITHUB_COLLABORATOR_ENTITY_TYPE,
+    _key: data.node_id,
+    login: data.login,
+    username: data.login,
+    displayName: data.name || data.login,
+    name: data.name || data.login,
+    mfaEnabled: undefined,
+    role: 'outside collaborator',
+    siteAdmin: false,
+  };
+  setRawData(userEntity, { name: 'default', rawData: data });
+  return userEntity;
+}
+
 export function toPullRequestEntity(
   data: PullsListResponseItem,
   commits?: PullsListCommitsResponseItem[],
@@ -157,8 +223,10 @@ export function toPullRequestEntity(
     : undefined;
 
   const approverLogins = approvals
-    ? flattenMatrix<string>(
-        aggregateProperties<string[]>('approverUsernames', approvals),
+    ? uniq(
+        flattenMatrix<string>(
+          aggregateProperties<string[]>('approverUsernames', approvals),
+        ),
       )
     : undefined;
   const userLogin: string = data.user ? data.user.login : '';
@@ -191,10 +259,10 @@ export function toPullRequestEntity(
     displayName: `${data.base.repo.name}/${data.number}`,
     accountLogin: data.base.repo.owner ? data.base.repo.owner.login : '',
     repository: data.base.repo.name,
-    //the type is hacked here because typing of data properties is controlled by a library call
-    //so I can't just say that data.number is a string
-    //here would be another way to solve it:
-    //id: JSON.stringify(data.number).replace(/\"/g, ''),
+    // the type is hacked here because typing of data properties is controlled by a library call
+    // so I can't just say that data.number is a string
+    // here would be another way to solve it:
+    // id: JSON.stringify(data.number).replace(/\"/g, ''),
     id: <string>(<unknown>data.number),
 
     name: data.title,
@@ -207,6 +275,7 @@ export function toPullRequestEntity(
 
     state: data.state,
     open: data.state === PRState.Open,
+    mergeCommitHash: data.merge_commit_sha,
     merged: (data.merged_at as any) !== null,
     declined: data.state === PRState.Closed && (data.merged_at as any) === null,
     approved,
@@ -223,6 +292,7 @@ export function toPullRequestEntity(
 
     createdOn: parseTimePropertyValue(data.created_at),
     updatedOn: parseTimePropertyValue(data.updated_at),
+    mergedOn: parseTimePropertyValue(data.merged_at),
 
     authorLogin: userLogin,
     author: authorUser
@@ -240,6 +310,46 @@ export function toPullRequestEntity(
         ? displayNamesFromLogins(approverLogins, usersByLogin)
         : approverLogins,
   };
-  setRawData(entity, { name: 'default', rawData: data });
+  const rawDataPropertiesToRemove = ['head', 'base']; // a few particularly large pieces of data that are repeated on every PR
+  setRawData(entity, {
+    name: 'default',
+    rawData: omit(data, rawDataPropertiesToRemove),
+  });
   return entity;
+}
+
+export function createRepoAllowsTeamRelationship(
+  repo: RepoEntity,
+  team: TeamEntity,
+  permission: string,
+): RepoTeamRelationship {
+  return {
+    _key: `${repo._key}|allows|${team._key}`,
+    _class: RelationshipClass.ALLOWS,
+    _type: GITHUB_REPO_TEAM_RELATIONSHIP_TYPE,
+    _fromEntityKey: repo._key,
+    _toEntityKey: team._key,
+    displayName: RelationshipClass.ALLOWS,
+    permission: permission,
+  };
+}
+
+export function createRepoAllowsUserRelationship(
+  repo: RepoEntity,
+  user: UserEntity,
+  permissions?: CollaboratorPermissions,
+): Relationship {
+  return {
+    _key: `${repo._key}|allows|${user._key}`,
+    _class: RelationshipClass.ALLOWS,
+    _type: GITHUB_REPO_USER_RELATIONSHIP_TYPE,
+    _fromEntityKey: repo._key,
+    _toEntityKey: user._key,
+    displayName: RelationshipClass.ALLOWS,
+    adminPermission: permissions?.admin,
+    maintainPermission: permissions?.maintain,
+    pushPermission: permissions?.push,
+    triagePermission: permissions?.triage,
+    pullPermission: permissions?.pull,
+  };
 }